jiff::civil

Module date

Structs

Date

A representation of a civil date in the Gregorian calendar.

DateArithmetic

Options for Date::checked_add and Date::checked_sub.

DateDifference

Options for Date::since and Date::until.

DateSeries

An iterator over periodic dates, created by Date::series.

DateWith

A builder for setting the fields on a Date.

Enums

DateWithDay 🔒

Encodes the “with day” option of DateWith.

DateWithYear 🔒

Encodes the “with year” option of DateWith.

Functions

month_add_one 🔒

Adds or subtracts sign from the given year/month.

month_add_overflowing 🔒

Adds the given span of months to the month given.