jiff::civil::date

Function month_add_overflowing

fn month_add_overflowing(month: i8, span: i32) -> (i8, i16)

Adds the given span of months to the month given.

If adding (or subtracting) would result in overflowing the month value, then the amount by which it overflowed, in units of years, is returned. For example, adding 14 months to the month 3 (March) will result in returning the month 5 (May) with 1 year of overflow.

Preconditions

Callers must ensure that span is in bounds for b::SpanMonths.