jiff::civil

Module datetime

Structs

DateTime

A representation of a civil datetime in the Gregorian calendar.

DateTimeArithmetic

Options for DateTime::checked_add and DateTime::checked_sub.

DateTimeDifference

Options for DateTime::since and DateTime::until.

DateTimeRound

Options for DateTime::round.

DateTimeSeries

An iterator over periodic datetimes, created by DateTime::series.

DateTimeWith

A builder for setting the fields on a DateTime.