jiff::civil

Struct DateTimeArithmetic

pub struct DateTimeArithmetic { /* private fields */ }

Options for DateTime::checked_add and DateTime::checked_sub.

This type provides a way to ergonomically add one of a few different duration types to a DateTime.

The main way to construct values of this type is with its From trait implementations:

From<Span> for DateTimeArithmetic adds (or subtracts) the given span to the receiver datetime.

From<SignedDuration> for DateTimeArithmetic adds (or subtracts) the given signed duration to the receiver datetime.

From<std::time::Duration> for DateTimeArithmetic adds (or subtracts) the given unsigned duration to the receiver datetime.

Example

use std::time::Duration;

use jiff::{civil::date, SignedDuration, ToSpan};

let dt = date(2024, 2, 29).at(0, 0, 0, 0);
assert_eq!(
    dt.checked_add(1.year())?,
    date(2025, 2, 28).at(0, 0, 0, 0),
);
assert_eq!(
    dt.checked_add(SignedDuration::from_hours(24))?,
    date(2024, 3, 1).at(0, 0, 0, 0),
);
assert_eq!(
    dt.checked_add(Duration::from_secs(24 * 60 * 60))?,
    date(2024, 3, 1).at(0, 0, 0, 0),
);