jiff

Module error

Modules

civil 🔒

duration 🔒

fmt 🔒

signed_duration 🔒

span 🔒

timestamp 🔒

tz 🔒

unit 🔒

util 🔒

zoned 🔒

Structs

AdhocError 🔒

A generic error message.

Error

An error that can occur in this crate.

ErrorInner 🔒

FilePathError 🔒

IOError 🔒

A std::io::Error.

Enums

CrateFeatureError 🔒

An error used whenever a failure is caused by a missing crate feature.

ErrorKind 🔒

The underlying kind of a Error.

Traits

ErrorContext 🔒

A trait for contextualizing error values.

IntoError 🔒

A simple trait to encapsulate automatic conversion to Error.