jiff::fmt::friendly

Module printer

Structs

DesignatorWriter 🔒

An abstraction for writing the “designator” variant of the friendly format.

Designators 🔒

A type that represents the designator choice.

FractionalPrinter 🔒

A printer for a fraction with an integer and fraction component.

SpanPrinter

A printer for Jiff’s “friendly” duration format.

Enums

Designator

Configuration for SpanPrinter::designator.

Direction

Configuration for SpanPrinter::direction.

DirectionSign 🔒

The sign to write and whether it should be a prefix or a suffix.

FractionalUnit

Configuration for SpanPrinter::fractional.

Spacing

Configuration for SpanPrinter::spacing.

Constants

MAX_SIGNED_DURATION_LEN 🔒

Defines the maximum possible length (in bytes) of a SignedDuration printed in the friendly format.

MAX_SPAN_LEN 🔒

Defines the maximum possible length (in bytes) of a Span printed in the friendly format.

MAX_UNSIGNED_DURATION_LEN 🔒

Defines the maximum possible length (in bytes) of a std::time::Duration printed in the friendly format.

MINS_PER_HOUR 🔒

NANOS_PER_HOUR 🔒

NANOS_PER_MICRO 🔒

NANOS_PER_MILLI 🔒

NANOS_PER_MIN 🔒

SECS_PER_HOUR 🔒

SECS_PER_MIN 🔒