jiff::fmt::friendly::printer

Struct DesignatorWriter

struct DesignatorWriter<'p, 'w, 'd> {
    printer: &'p SpanPrinter,
    bbuf: &'w mut BorrowedBuffer<'d>,
    sign: Option<DirectionSign>,
    padding: u8,
    precision: Option<u8>,
    written_non_zero_unit: bool,
}

An abstraction for writing the “designator” variant of the friendly format.

This takes care of computing some initial state and keeping track of some mutable state that influences printing. For example, whether to write a delimiter or not (one should only come after a unit that has been written).

Fields
printer: &'p SpanPrinterbbuf: &'w mut BorrowedBuffer<'d>sign: Option<DirectionSign>padding: u8precision: Option<u8>written_non_zero_unit: bool
Implementations

impl<'p, 'w, 'd> DesignatorWriter<'p, 'w, 'd>

fn new( printer: &'p SpanPrinter, bbuf: &'w mut BorrowedBuffer<'d>, has_calendar: bool, signum: i8, ) -> DesignatorWriter<'p, 'w, 'd>

fn maybe_write_prefix_sign(&mut self)

fn maybe_write_suffix_sign(&mut self)

fn maybe_write_zero(&mut self)

fn write(&mut self, unit: Unit, value: u64)

fn write_fractional_duration( &mut self, unit: FractionalUnit, duration: &Duration, )

fn finish_preceding(&mut self)