jiff::fmt

Module offset

This module provides facilities for parsing time zone offsets.

The parsing here follows primarily from RFC 3339 and ISO 8601, but also from Temporal’s hybrid grammar.

Structs

Numeric 🔒

A numeric representation of a UTC offset.

ParsedOffset 🔒

An offset that has been parsed from a datetime string.

Parser 🔒

A parser for UTC offsets.

Enums

Colon 🔒

How to handle parsing of colons in a time zone offset.

ParsedOffsetKind 🔒

The kind of a parsed offset.

Type Aliases

ParsedOffsetHours 🔒

ParsedOffsetMinutes 🔒

ParsedOffsetSeconds 🔒