jiff::fmt::offset

Struct Numeric

struct Numeric {
    sign: ri8<-1, 1>,
    hours: ri8<0, { t::SpanZoneOffsetHours::MAX }>,
    minutes: Option<ri8<0, { t::SpanZoneOffsetMinutes::MAX }>>,
    seconds: Option<ri8<0, { t::SpanZoneOffsetSeconds::MAX }>>,
    nanoseconds: Option<ri32<0, { _ }>>,
}

A numeric representation of a UTC offset.

Fields
sign: ri8<-1, 1>

The sign that was parsed from the numeric UTC offset. This is always either 1 or -1, never 0.

hours: ri8<0, { t::SpanZoneOffsetHours::MAX }>

The hours component. This is non-optional because every UTC offset must have at least hours.

minutes: Option<ri8<0, { t::SpanZoneOffsetMinutes::MAX }>>

The minutes component.

seconds: Option<ri8<0, { t::SpanZoneOffsetSeconds::MAX }>>

The seconds component. This is only possible when subminute resolution is enabled.

nanoseconds: Option<ri32<0, { _ }>>

The nanoseconds fractional component. This is only possible when subminute resolution is enabled.

Implementations

impl Numeric

fn to_offset(&self) -> Result<Offset, Error>

Convert a parsed numeric offset into a Jiff offset.

This does rounding based on the fractional nanosecond part. As a result, if the parsed value would be rounded to a value not in bounds for a Jiff offset, this returns an error.