jiff::fmt::strtime::format

Function write_offset

fn write_offset(
    offset: Offset,
    colon: bool,
    minute: bool,
    second: bool,
    wtr: &mut BorrowedWriter<'_, '_, '_>,
) -> Result<(), Error>

Writes the given time zone offset to the writer.

When colon is true, the hour, minute and optional second components are delimited by a colon. Otherwise, no delimiter is used.

When minute is true, the minute component is always printed. When false, the minute component is only printed when it is non-zero (or if the second component is non-zero).

When second is true, the second component is always printed. When false, the second component is only printed when it is non-zero.