jiff::fmt::strtime

Struct Extension

pub struct Extension {
    flag: Option<Flag>,
    width: Option<u8>,
    colons: u8,
}

These are “extensions” to the standard strftime conversion specifiers.

This type represents which flags and/or padding were provided with a specifier. For example, %_3d uses 3 spaces of padding.

Currently, this type provides no structured introspection facilities. It is exported and available only via implementations of the Custom trait for reasons of semver compatible API evolution. If you have use cases for introspecting this type, please open an issue.

Fields
flag: Option<Flag>width: Option<u8>colons: u8
Implementations

impl Extension

fn parse_number<'i>( &self, default_pad_width: usize, default_flag: Flag, inp: &'i [u8], ) -> Result<(i64, &'i [u8]), Error>

Parse an integer with the given default padding and flag settings.

The default padding is usually 2 (4 for %Y) and the default flag is usually Flag::PadZero (there are no cases where the default flag is different at time of writing). But both the padding and the flag can be overridden by the settings on this extension.

Generally speaking, parsing ignores everything in an extension except for padding. When padding is set, then parsing will limit itself to a number of digits equal to the greater of the default padding size or the configured padding size. This permits %Y%m%d to parse 20240730 successfully, for example.

The remaining input is returned. This returns an error if the given input is empty.

impl Extension

fn write_str( &self, default: Case, string: &str, wtr: &mut BorrowedWriter<'_, '_, '_>, ) -> Result<(), Error>

Writes the given string using the default case rule provided, unless an option in this extension config overrides the default case.

fn write_str_cold( &self, default: Case, string: &str, wtr: &mut BorrowedWriter<'_, '_, '_>, ) -> Result<(), Error>

fn write_int( &self, pad_byte: u8, pad_width: u8, number: impl Into<i64>, wtr: &mut BorrowedWriter<'_, '_, '_>, ) -> Result<(), Error>

Writes the given integer using the given padding width and byte, unless an option in this extension config overrides a default setting.

fn write_negative_int( pad_byte: u8, pad_width: u8, number: u64, wtr: &mut BorrowedWriter<'_, '_, '_>, ) -> Result<(), Error>

Writes a negative integer with explicit width, where sign is part of total width.

impl Extension

fn parse_flag<'i>(fmt: &'i [u8]) -> Result<(Option<Flag>, &'i [u8]), Error>

Parses an optional directive flag from the beginning of fmt. This assumes fmt is not empty and guarantees that the return unconsumed slice is also non-empty.

fn parse_width<'i>(fmt: &'i [u8]) -> Result<(Option<u8>, &'i [u8]), Error>

Parses an optional width that comes after a (possibly absent) flag and before the specifier directive itself. And if a width is parsed, the slice returned does not contain it. (If that slice is empty, then an error is returned.)

Note that this is also used to parse precision settings for %f and %.f. In the former case, the width is just re-interpreted as a precision setting. In the latter case, something like %5.9f is technically valid, but the 5 is ignored.

fn parse_colons<'i>(fmt: &'i [u8]) -> Result<(u8, &'i [u8]), Error>

Parses an optional number of colons.

This is meant to be used immediately before the conversion specifier (after the flag and width has been parsed).

This supports parsing up to 3 colons. The colons are used in some cases for alternate specifiers. e.g., %:Q or %:::z.