jiff::fmt::temporal

Module pieces

Structs

Pieces

A low level representation of a parsed Temporal ISO 8601 datetime string.

PiecesNumericOffset

A specific numeric offset, including the sign of the offset, for use with Pieces.

TimeZoneAnnotation

An RFC 9557 time zone annotation, for use with Pieces.

TimeZoneAnnotationName

A time zone annotation parsed from a datetime string.

Enums

PiecesOffset

An offset parsed from a Temporal ISO 8601 datetime string, for use with Pieces.

TimeZoneAnnotationKind

The kind of time zone found in an RFC 9557 timestamp, for use with Pieces.