jiff::fmt::temporal::printer

Struct SpanPrinter

pub(super) struct SpanPrinter {
    designators: &'static Designators,
}

A printer for Temporal spans.

Note that in Temporal, a “span” is called a “duration.”

Fields
designators: &'static Designators

The designators to use.

Implementations

impl SpanPrinter

pub(super) const fn new() -> SpanPrinter

Create a new Temporal span printer with the default configuration.

pub(super) const fn lowercase(self, yes: bool) -> SpanPrinter

Use lowercase for unit designator labels.

By default, unit designator labels are written in uppercase.

pub(super) fn print_span<W: Write>( &self, span: &Span, wtr: W, ) -> Result<(), Error>

Print the given span to the writer given.

This only returns an error when the given writer returns an error.

fn print_span_impl(&self, span: &Span, bbuf: &mut BorrowedBuffer<'_>)

pub(super) fn print_signed_duration<W: Write>( &self, dur: &SignedDuration, wtr: W, ) -> Result<(), Error>

Print the given signed duration to the writer given.

This only returns an error when the given writer returns an error.

pub(super) fn print_unsigned_duration<W: Write>( &self, dur: &Duration, wtr: W, ) -> Result<(), Error>

Print the given unsigned duration to the writer given.

This only returns an error when the given writer returns an error.

fn print_unsigned_duration_impl( &self, dur: &Duration, bbuf: &mut BorrowedBuffer<'_>, )

fn label(&self, unit: Unit) -> u8

Converts the ASCII uppercase unit designator label to lowercase if this printer is configured to use lowercase. Otherwise the label is returned unchanged.