jiff::fmt

Module util

Structs

DurationUnits 🔒

A container for holding a partially parsed duration.

Functions

duration_unit_value 🔒

Returns the given parsed value, interpreted as the given unit, as a SignedDuration.

fractional_duration 🔒

Converts the fraction of the given unit to a signed duration.

fractional_time_to_duration 🔒

Like fractional_time_to_span, but just converts the fraction of the given unit to a signed duration.

fractional_time_to_span 🔒

This routine returns a span based on the given unit and value with fractional time applied to it.

parse_temporal_fraction 🔒

Parses an optional fractional number from the start of input.