jiff

Module span

Macros

impl_to_span 🔒

Structs

Nudge 🔒

The result of a span rounding strategy. There are three:

RelativeCivil 🔒

A wrapper around a civil datetime and a timestamp corresponding to that civil datetime in UTC.

RelativeSpan 🔒

A balanced span between a range of civil or zoned datetimes.

RelativeZoned 🔒

A simple wrapper around a possibly borrowed Zoned.

Span

A span of time represented via a mixture of calendar and clock units.

SpanArithmetic

Options for Span::checked_add and Span::checked_sub.

SpanCompare

Options for Span::compare.

SpanFieldwise

A wrapper for Span that implements the Hash, Eq and PartialEq traits.

SpanRelativeTo

A relative datetime for use with Span APIs.

SpanRound

Options for Span::round.

SpanTotal

Options for Span::total.

UnitSet 🔒

A bit set that keeps track of all non-zero units on a Span.

Enums

Relative 🔒

An internal abstraction for managing a relative datetime for use in some Span APIs.

RelativeSpanKind 🔒

A civil or zoned datetime range of time.

SpanRelativeToKind 🔒

Unit

A way to refer to a single calendar or clock unit.

Traits

ToSpan

A trait for enabling concise literals for creating Span values.

Functions

parse_iso_or_friendly 🔒

A common parsing function that works in bytes.

requires_relative_date_err 🔒

round_span_invariant 🔒

Rounds a span consisting of only invariant units.

unit_start_and_end 🔒

Returns the nanosecond timestamps of relative + span and relative + {amount of unit} + span.