jiff::span

Struct RelativeZoned

struct RelativeZoned<'a> {
    zoned: DumbCow<'a, Zoned>,
}

A simple wrapper around a possibly borrowed Zoned.

Fields
zoned: DumbCow<'a, Zoned>
Implementations

impl<'a> RelativeZoned<'a>

fn checked_add(&self, span: Span) -> Result<RelativeZoned<'static>, Error>

Returns the result of Zoned::checked_add.

Errors

Returns an error in the same cases as Zoned::checked_add. That is, when adding the span to this zoned datetime would overflow.

fn checked_add_duration( &self, duration: SignedDuration, ) -> Result<RelativeZoned<'static>, Error>

Returns the result of Zoned::checked_add with an absolute duration.

Errors

Returns an error in the same cases as Zoned::checked_add. That is, when adding the span to this zoned datetime would overflow.

fn until(&self, largest: Unit, other: &RelativeZoned<'a>) -> Result<Span, Error>

Returns the result of Zoned::until.

Errors

Returns an error in the same cases as Zoned::until. That is, when the span for the given largest unit cannot be represented. This can generally only happen when largest is Unit::Nanosecond and the span cannot be represented as a 64-bit integer of nanoseconds.

fn borrowed(&'a self) -> RelativeZoned<'a>

Returns the borrowed version of self; useful when you need to convert &RelativeZoned into RelativeZoned without cloning anything.