jiff

Struct SpanArithmetic

pub struct SpanArithmetic<'a> { /* private fields */ }

Options for Span::checked_add and Span::checked_sub.

This type provides a way to ergonomically add two spans with an optional relative datetime. Namely, a relative datetime is only needed when at least one of the two spans being added (or subtracted) has a non-zero calendar unit (years, months, weeks or days). Otherwise, an error will be returned.

Callers may use SpanArithmetic::days_are_24_hours to opt into 24-hour invariant days (and 7-day weeks) without providing a relative datetime.

The main way to construct values of this type is with its From trait implementations:

From<Span> for SpanArithmetic adds (or subtracts) the given span to the receiver in Span::checked_add (or Span::checked_sub).

From<(Span, civil::Date)> for SpanArithmetic adds (or subtracts) the given span to the receiver in Span::checked_add (or Span::checked_sub), relative to the given date. There are also From implementations for civil::DateTime, Zoned and SpanRelativeTo.

Example

use jiff::ToSpan;

assert_eq!(
    1.hour().checked_add(30.minutes())?,
    1.hour().minutes(30).fieldwise(),
);

Implementations

impl<'a> SpanArithmetic<'a>

pub fn days_are_24_hours(self) -> SpanArithmetic<'a>

This is a convenience function for setting the relative option on this configuration to SpanRelativeTo::days_are_24_hours.

Example

When doing arithmetic on spans involving days, either a relative datetime must be provided, or a special assertion opting into 24-hour days is required. Otherwise, you get an error.

use jiff::{SpanArithmetic, ToSpan};

let span1 = 2.days().hours(12);
let span2 = 12.hours();
// No relative date provided, which results in an error.
assert_eq!(
    span1.checked_add(span2).unwrap_err().to_string(),
    "using unit 'day' in a span or configuration requires that \
     either a relative reference time be given or \
     `SpanRelativeTo::days_are_24_hours()` is used to indicate \
     invariant 24-hour days, but neither were provided",
);
let sum = span1.checked_add(
    SpanArithmetic::from(span2).days_are_24_hours(),
)?;
assert_eq!(sum, 3.days().fieldwise());