jiff

Struct ZonedRound

pub struct ZonedRound { /* private fields */ }

Options for Zoned::round.

This type provides a way to configure the rounding of a zoned datetime. In particular, Zoned::round accepts anything that implements the Into<ZonedRound> trait. There are some trait implementations that therefore make calling Zoned::round in some common cases more ergonomic:

From<Unit> for ZonedRound will construct a rounding configuration that rounds to the unit given. Specifically, ZonedRound::new().smallest(unit).

From<(Unit, i64)> for ZonedRound is like the one above, but also specifies the rounding increment for ZonedRound::increment.

Note that in the default configuration, no rounding occurs.

Example

This example shows how to round a zoned datetime to the nearest second:

use jiff::{civil::date, Unit, Zoned};

let zdt: Zoned = "2024-06-20 16:24:59.5[America/New_York]".parse()?;
assert_eq!(
    zdt.round(Unit::Second)?,
    // The second rounds up and causes minutes to increase.
    date(2024, 6, 20).at(16, 25, 0, 0).in_tz("America/New_York")?,
);

The above makes use of the fact that Unit implements Into<ZonedRound>. If you want to change the rounding mode to, say, truncation, then youโll need to construct a ZonedRound explicitly since there are no convenience Into trait implementations for RoundMode.

use jiff::{civil::date, RoundMode, Unit, Zoned, ZonedRound};

let zdt: Zoned = "2024-06-20 16:24:59.5[America/New_York]".parse()?;
assert_eq!(
    zdt.round(
        ZonedRound::new().smallest(Unit::Second).mode(RoundMode::Trunc),
    )?,
    // The second just gets truncated as if it wasn't there.
    date(2024, 6, 20).at(16, 24, 59, 0).in_tz("America/New_York")?,
);

Implementations

impl ZonedRound

pub fn new() -> ZonedRound

Create a new default configuration for rounding a Zoned.

pub fn smallest(self, unit: Unit) -> ZonedRound

Set the smallest units allowed in the zoned datetime returned after rounding.

Any units below the smallest configured unit will be used, along with the rounding increment and rounding mode, to determine the value of the smallest unit. For example, when rounding 2024-06-20T03:25:30[America/New_York] to the nearest minute, the 30 second unit will result in rounding the minute unit of 25 up to 26 and zeroing out everything below minutes.

This defaults to Unit::Nanosecond.

Errors

The smallest units must be no greater than Unit::Day. And when the smallest unit is Unit::Day, the rounding increment must be equal to 1. Otherwise an error will be returned from Zoned::round.

Example

use jiff::{civil::date, Unit, ZonedRound};

let zdt = date(2024, 6, 20).at(3, 25, 30, 0).in_tz("America/New_York")?;
assert_eq!(
    zdt.round(ZonedRound::new().smallest(Unit::Minute))?,
    date(2024, 6, 20).at(3, 26, 0, 0).in_tz("America/New_York")?,
);
// Or, utilize the `From<Unit> for ZonedRound` impl:
assert_eq!(
    zdt.round(Unit::Minute)?,
    date(2024, 6, 20).at(3, 26, 0, 0).in_tz("America/New_York")?,
);

pub fn mode(self, mode: RoundMode) -> ZonedRound

Set the rounding mode.

This defaults to RoundMode::HalfExpand, which rounds away from zero. It matches the kind of rounding you might have been taught in school.

Example

This shows how to always round zoned datetimes up towards positive infinity.

use jiff::{civil::date, RoundMode, Unit, Zoned, ZonedRound};

let zdt: Zoned = "2024-06-20 03:25:01[America/New_York]".parse()?;
assert_eq!(
    zdt.round(
        ZonedRound::new()
            .smallest(Unit::Minute)
            .mode(RoundMode::Ceil),
    )?,
    date(2024, 6, 20).at(3, 26, 0, 0).in_tz("America/New_York")?,
);

pub fn increment(self, increment: i64) -> ZonedRound

Set the rounding increment for the smallest unit.

The default value is 1. Other values permit rounding the smallest unit to the nearest integer increment specified. For example, if the smallest unit is set to Unit::Minute, then a rounding increment of 30 would result in rounding in increments of a half hour. That is, the only minute value that could result would be 0 or 30.

Errors

When the smallest unit is Unit::Day, then the rounding increment must be 1 or else Zoned::round will return an error.

For other units, the rounding increment must divide evenly into the next highest unit above the smallest unit set. The rounding increment must also not be equal to the next highest unit. For example, if the smallest unit is Unit::Nanosecond, then some of the valid values for the rounding increment are 1, 2, 4, 5, 100 and 500. Namely, any integer that divides evenly into 1,000 nanoseconds since there are 1,000 nanoseconds in the next highest unit (microseconds).

Example

This example shows how to round a zoned datetime to the nearest 10 minute increment.

use jiff::{civil::date, RoundMode, Unit, Zoned, ZonedRound};

let zdt: Zoned = "2024-06-20 03:24:59[America/New_York]".parse()?;
assert_eq!(
    zdt.round((Unit::Minute, 10))?,
    date(2024, 6, 20).at(3, 20, 0, 0).in_tz("America/New_York")?,
);