jiff

Module timestamp

Structs

Timestamp

An instant in time represented as the number of nanoseconds since the Unix epoch.

TimestampArithmetic

Options for Timestamp::checked_add and Timestamp::checked_sub.

TimestampDifference

Options for Timestamp::since and Timestamp::until.

TimestampDisplayWithOffset

A type for formatting a Timestamp with a specific offset.

TimestampRound

Options for Timestamp::round.

TimestampSeries

An iterator over periodic timestamps, created by Timestamp::series.