jiff::tz::timezone::repr

Module polyfill

This is a polyfill for a small subset of std’s strict provenance APIs.

The strict provenance APIs in core were stabilized in Rust 1.84, but it will likely be a while before Jiff can use them. (At time of writing, 2025-02-24, Jiff’s MSRV is Rust 1.70.)

Traits

StrictProvenancePolyfill 🔒

Functions

without_provenance 🔒