jiff::util

Module b

A module for constants and various base utilities.

This module is a work-in-progress that may lead to helping us move off of ranged integers. I’m not quite sure where this will go.

Macros

define_bounds 🔒

This macro writes out the boiler plate to define a boundary type.

define_one_boundary_type 🔒

Writes the definition of a single boundary type.

Structs

Century 🔒

CivilDayNanosecond 🔒

CivilDaySecond 🔒

Day 🔒

DayOfYear 🔒

Hour 🔒

Hour12 🔒

ISOWeek 🔒

ISOYear 🔒

Increment 🔒

Increment32 🔒

LeapSecond 🔒

Microsecond 🔒

Millisecond 🔒

Minute 🔒

Month 🔒

Nanosecond 🔒

NthWeekday 🔒

OffsetHours 🔒

OffsetMinutes 🔒

OffsetSeconds 🔒

OffsetTotalSeconds 🔒

Second 🔒

SignedDurationSeconds 🔒

SignedSubsecNanosecond 🔒

SpanDays 🔒

SpanHours 🔒

SpanMicroseconds 🔒

SpanMilliseconds 🔒

SpanMinutes 🔒

SpanMonths 🔒

SpanMultiple 🔒

SpanNanoseconds 🔒

SpanSeconds 🔒

SpanWeeks 🔒

SpanYears 🔒

SubsecNanosecond 🔒

UnixEpochDays 🔒

UnixEpochMicroseconds 🔒

UnixEpochMilliseconds 🔒

UnixEpochSeconds 🔒

WeekNum 🔒

WeekdayMondayOne 🔒

WeekdayMondayZero 🔒

WeekdaySundayOne 🔒

WeekdaySundayZero 🔒

Year 🔒

YearBCE 🔒

YearCE 🔒

YearTwoDigit 🔒

ZonedDayNanoseconds 🔒

ZonedDaySeconds 🔒

Enums

BoundsError 🔒

An error that indicates a value is out of its intended range.

SpecialBoundsError 🔒

Like BoundsError, but maintained manually.