jiff::util::fs

Function last_modified_from_path

pub(crate) fn last_modified_from_path(path: &Path) -> Option<Timestamp>

Returns the last modified time for the given file path as a Jiff timestamp.

If there was a problem accessing the last modified time or if it could not fit in a Jiff timestamp, then a warning message is logged and None is returned.