jiff::util

Module t

Structs

Constant 🔒

A constant value for use in arithmetic in this crate.

Constants

DAYS_PER_CIVIL_WEEK 🔒

The number of days in a week.

HOURS_PER_CIVIL_DAY 🔒

The number of whole hours in one day.

MICROS_PER_CIVIL_DAY 🔒

The number of microseconds in a civil day.

MICROS_PER_MILLI 🔒

The number of microseconds in a single millisecond.

MICROS_PER_SECOND 🔒

The number of microseconds in a single second.

MILLIS_PER_CIVIL_DAY 🔒

The number of microseconds in a civil day.

MILLIS_PER_SECOND 🔒

The number of milliseconds in a single second.

MINUTES_PER_CIVIL_DAY 🔒

The number of minutes in a civil day.

MINUTES_PER_HOUR 🔒

The number of minutes in an hour.

MONTHS_PER_YEAR 🔒

The number of months in a year.

NANOS_PER_CIVIL_DAY 🔒

The number of nanoseconds in a civil day.

NANOS_PER_CIVIL_WEEK 🔒

The number of nanoseconds in a civil week.

NANOS_PER_HOUR 🔒

The number of nanoseconds in a single hour.

NANOS_PER_MICRO 🔒

The number of nanoseconds in a single microsecond.

NANOS_PER_MILLI 🔒

The number of nanoseconds in a single millisecond.

NANOS_PER_MINUTE 🔒

The number of nanoseconds in a single minute.

NANOS_PER_SECOND 🔒

The number of nanoseconds in a single second.

SECONDS_PER_CIVIL_DAY 🔒

The number of seconds in a civil day.

SECONDS_PER_CIVIL_WEEK 🔒

The number of seconds in a civil week.

SECONDS_PER_HOUR 🔒

The number of seconds in a single hour.

SECONDS_PER_MINUTE 🔒

The number of seconds in a single minute.

SPAN_CIVIL_SECONDS 🔒

The maximum number of seconds that can be expressed with a span.

SPAN_ZONE_OFFSET_TOTAL_SECONDS 🔒

The max number of seconds that can be expressed in a time zone offset.

Functions

C 🔒

A short-hand creating a generic Constant value as a ranged integer.

C128 🔒

next_multiple_of 🔒

Computes the next multiple of rhs that is greater than or equal to lhs.

sign 🔒

Type Aliases

CivilDayNanosecond 🔒

A range representing each possible nanosecond in a single civil day.

CivilDaySecond 🔒

A range representing each possible second in a single civil day.

Day 🔒

The range of possible day values.

DayOfYear 🔒

FractionalMicrosecond 🔒

The range of allowable fractional microseconds.

FractionalMillisecond 🔒

The range of allowable fractional milliseconds.

FractionalNanosecond 🔒

The range of allowable fractional nanoseconds.

Hour 🔒

The range of possible hour values.

ISOWeek 🔒

ISOYear 🔒

Microsecond 🔒

The range of possible microsecond values.

Millisecond 🔒

The range of possible millisecond values.

Minute 🔒

The range of possible minute values.

Month 🔒

The range of possible month values.

Nanosecond 🔒

The range of possible nanosecond values.

NoUnits 🔒

A type alias for a ranged integer with no units.

NoUnits16 🔒

A type alias for a ranged 16-bit integer with no units.

NoUnits32 🔒

A type alias for a ranged 32-bit integer with no units.

NoUnits96 🔒

A type alias for a ranged 96-bit integer with no units.

NoUnits128 🔒

A type alias for a ranged 128-bit integer with no units.

Second 🔒

The range of possible second values not accounting for leap seconds.

Sign 🔒

A type alias for the sign of a number.

SpanDays 🔒

A range of the allowed number of days.

SpanHours 🔒

A range of the allowed number of hours.

SpanMicroseconds 🔒

A range of the allowed number of microseconds.

SpanMilliseconds 🔒

A range of the allowed number of milliseconds.

SpanMinutes 🔒

A range of the allowed number of minutes.

SpanMonths 🔒

A precise min/max of the allowed range of a duration in months.

SpanNanoseconds 🔒

A range of the allowed number of nanoseconds.

SpanSeconds 🔒

A range of the allowed number of seconds.

SpanWeeks 🔒

A range of the allowed number of weeks.

SpanYears 🔒

A precise min/max of the allowed range of a duration in years.

SpanZoneOffset 🔒

The span of seconds permitted for expressing the offset of a time zone.

SpanZoneOffsetHours 🔒

The number of hours allowed in a time zone offset.

SpanZoneOffsetMinutes 🔒

The number of minutes allowed in a time zone offset.

SpanZoneOffsetSeconds 🔒

The number of seconds allowed in a time zone offset.

SubsecNanosecond 🔒

The range of possible nanosecond values.

UnixEpochDay 🔒

The number of days from the Unix epoch for the Gregorian calendar.

UnixMicroseconds 🔒

Like UnixSeconds, but expressed in units of microseconds.

UnixMilliseconds 🔒

Like UnixSeconds, but expressed in units of milliseconds.

UnixNanoseconds 🔒

Like UnixSeconds, but expressed in units of nanoseconds.

UnixSeconds 🔒

The range of Unix seconds supported by Jiff.

WeekNum 🔒

WeekdayOne 🔒

The range of a weekday, offset from one.

WeekdayZero 🔒

The range of a weekday, offset from zero.

Year 🔒

The range of years supported by jiff.

YearBCE 🔒

The range of BCE years supported by jiff.

YearCE 🔒

The range of CE years supported by jiff.

ZonedDayNanoseconds 🔒

The number of nanoseconds permitted in a single day.

ZonedDaySeconds 🔒

The number of seconds permitted in a single day.