jiff

Module zoned

Structs

Zoned

A time zone aware instant in time.

ZonedArithmetic

Options for Timestamp::checked_add and Timestamp::checked_sub.

ZonedDifference

Options for Zoned::since and Zoned::until.

ZonedInner 🔒

The representation of a Zoned.

ZonedRound

Options for Zoned::round.

ZonedSeries

An iterator over periodic zoned datetimes, created by Zoned::series.

ZonedWith

A builder for setting the fields on a Zoned.