jiff

Module error

Modules

civil ğŸ”’

duration ğŸ”’

fmt ğŸ”’

signed_duration ğŸ”’

span ğŸ”’

timestamp ğŸ”’

tz ğŸ”’

util ğŸ”’

zoned ğŸ”’

Structs

AdhocError ğŸ”’

A generic error message.

Error

An error that can occur in this crate.

ErrorInner ğŸ”’

FilePathError ğŸ”’

IOError ğŸ”’

A std::io::Error.

RangeError ğŸ”’

An error that occurs when an input value is out of bounds.

SlimRangeError ğŸ”’

A slim error that occurs when an input value is out of bounds.

Enums

CrateFeatureError ğŸ”’

An error used whenever a failure is caused by a missing crate feature.

ErrorKind ğŸ”’

The underlying kind of a Error.

Traits

ErrorContext ğŸ”’

A trait for contextualizing error values.

IntoError ğŸ”’

A simple trait to encapsulate automatic conversion to Error.