jiff::fmt

Module buffer

Structs

ArrayBuffer 🔒

An uninitialized slice of bytes of fixed size.

BorrowedBuffer 🔒

A borrowed buffer for writing into an uninitialized slice of bytes.

BorrowedWriter 🔒

A buffering abstraction on top of BorrowedBuffer.

Constants

BROAD_MINIMUM_BUFFER_LEN 🔒

The minimum buffer length required for any of BorrowedBuffer’s integer formatting APIs to work.

MAX_CAPACITY 🔒

MAX_INTEGER_LEN 🔒

MAX_PRECISION 🔒

RADIX_100_SPACE 🔒

All integers in the range 0..=99, space padded.

RADIX_100_ZERO 🔒

All integers in the range 0..=99, zero padded.

Functions

digits 🔒

Returns the number of digits in the decimal representation of n.