jiff::fmt::strtime

Module parse

Structs

Parser 🔒

Functions

parse_ampm 🔒

Like parse_choice, but specialized for AM/PM.

parse_choice 🔒

Parses the input such that, on success, the index of the first matching choice (via ASCII case insensitive comparisons) is returned, along with any remaining unparsed input.

parse_iana 🔒

parse_iana_component 🔒

Parses a single IANA name component. That is, the thing that leads all IANA time zone identifiers and the thing that must always come after a /. This returns an error if no component could be found.

parse_month_name_abbrev 🔒

Like parse_choice, but specialized for month name abbreviation.

parse_optional_sign 🔒

Parses an optional sign from the beginning of the input. If one isn’t found, then the sign returned is positive.

parse_weekday_abbrev 🔒

Like parse_choice, but specialized for weekday abbreviation.